backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.1
django-cors-headers>=4.4
celery>=5.4
redis>=5.0
python-dateutil>=2.9
gunicorn>=22.0
pytest>=8.3
pytest-django>=4.9
pytest-cov>=5.0
